+++ docs/changelog.rst
@@ -4,8 +4,9 @@ Changelog
 0.0.17 (2020-04-03)
 ~~~~~~~~~~~~~~~~~~~
 
--  :class:`~ocdsextensionregistry.extension.Extension`: Add :meth:`~ocdsextensionregistry.extension.Extension.__repr__`.
+-  :class:`~ocdsextensionregistry.extension_version.ExtensionVersion`: :meth:`~ocdsextensionregistry.extension_version.ExtensionVersion.remote` raises :exception:`~ocdsextensionregistry.exceptions.DoesNotExist` instead of :exception:`KeyError` if a file does not exist.
 -  :class:`~ocdsextensionregistry.extension_version.ExtensionVersion`: Add :meth:`~ocdsextensionregistry.extension_version.ExtensionVersion.__repr__`.
+-  :class:`~ocdsextensionregistry.extension.Extension`: Add :meth:`~ocdsextensionregistry.extension.Extension.__repr__`.
 
 0.0.16 (2019-11-20)
 ~~~~~~~~~~~~~~~~~~~